refactor: move DTOs, controllers, utils, and interfaces to dedicated directories
- Moved DTOs to `dto/` directory
- Moved controllers to `controllers/` directory
- Moved utility functions to `utils/` directory
- Moved interfaces to `interfaces/` directory
- Updated imports across multiple files to reflect the new structure
