+++ jest.config.js
@@ -10,5 +10,10 @@ module.exports = {
     '^@routes/(.*)$': '<rootDir>/src/routes/$1',
     '^@useCases/(.*)$': '<rootDir>/src/useCases/$1',
     '^@test/(.*)$': '<rootDir>/src/test/$1',
+    '^@enums/(.*)$': '<rootDir>/src/enums/$1',
+    '^@interfaces/(.*)$': '<rootDir>/src/interfaces/$1',
+    '^@utils/(.*)$': '<rootDir>/src/utils/$1',
+    '^@dto/(.*)$': '<rootDir>/src/dto/$1',
+    '^@controllers/(.*)$': '<rootDir>/src/controllers/$1',
   }
 };
