+++ src/enums/BmiCategoryEnum.ts
