+++ src/interfaces/entities/IPokemon.ts
@@ -1,0 +1,9 @@
+import { BmiCategoryEnum } from "@enums/BmiCategoryEnum";
+
+export interface IPokemon {
+  name: string;
+  height: number;
+  weight: number;
+  bmi?: number;
+  category?: BmiCategoryEnum;
+}
