+++ src/interfaces/services/IGetPokemonByName.ts
