+++ src/interfaces/services/IListPokemons.ts
