+++ src/test/controllers/GetPokemonByNameController.test.ts
@@ -3,13 +3,13 @@ dotenv.config();
 
 import request from 'supertest';
 import express, { Express, Request, Response } from 'express';
-import { BmiCalculator } from "@entities/BmiCalculator";
-import { GetPokemonByNameService } from "@services/GetPokemonByName/GetPokemonByNameService";
-import { GetPokemonByNameUseCase } from "@useCases/GetPokemonByName/GetPokemonByNameUseCase";
-import { GetPokemonByNameController } from "@useCases/GetPokemonByName/GetPokemonByNameController";
-import { BmiClassifier } from "@entities/BmiClassifier";
-import { HeightConverter } from "@entities/HeightConverter";
-import { WeightConverter } from "@entities/WeightConverter";
+import { BmiCalculator } from "@utils/BmiCalculator";
+import { GetPokemonByNameService } from "@services/GetPokemonByNameService";
+import { GetPokemonByNameUseCase } from "@useCases/GetPokemonByNameUseCase";
+import { GetPokemonByNameController } from "@controllers/GetPokemonByNameController";
+import { BmiClassifier } from "@utils/BmiClassifier";
+import { HeightConverter } from "@utils/HeightConverter";
+import { WeightConverter } from "@utils/WeightConverter";
 
 const app: Express = express();
 app.use(express.json());
