+++ src/test/services/GetPokemonByNameService.test.ts
@@ -1,11 +1,11 @@
 import dotenv from "dotenv";
 dotenv.config();
 
-import { BmiCalculator } from "@entities/BmiCalculator";
-import { GetPokemonByNameService } from "@services/GetPokemonByName/GetPokemonByNameService";
-import { BmiClassifier } from "@entities/BmiClassifier";
-import { HeightConverter } from "@entities/HeightConverter";
-import { WeightConverter } from "@entities/WeightConverter";
+import { BmiCalculator } from "@utils/BmiCalculator";
+import { GetPokemonByNameService } from "@services/GetPokemonByNameService";
+import { BmiClassifier } from "@utils/BmiClassifier";
+import { HeightConverter } from "@utils/HeightConverter";
+import { WeightConverter } from "@utils/WeightConverter";
 
 describe('GetPokemonByName', () => {
   const bmiCalculator = new BmiCalculator();
