+++ src/test/useCases/ListPokemonsUseCase.test.ts
@@ -1,12 +1,12 @@
 import dotenv from "dotenv";
 dotenv.config();
 
-import { BmiCalculator } from "@entities/BmiCalculator";
-import { ListPokemonsService } from "@services/ListPokemons/ListPokemonsService";
-import { ListPokemonsUseCase } from "@useCases/ListPokemons/ListPokemonsUseCase";
-import { BmiClassifier } from "@entities/BmiClassifier";
-import { HeightConverter } from "@entities/HeightConverter";
-import { WeightConverter } from "@entities/WeightConverter";
+import { BmiCalculator } from "@utils/BmiCalculator";
+import { ListPokemonsService } from "@services/ListPokemonsService";
+import { ListPokemonsUseCase } from "@useCases/ListPokemonsUseCase";
+import { BmiClassifier } from "@utils/BmiClassifier";
+import { HeightConverter } from "@utils/HeightConverter";
+import { WeightConverter } from "@utils/WeightConverter";
 
 describe('ListPokemonsUseCase', () => {
   const bmiCalculator = new BmiCalculator();
