+++ src/utils/BmiCalculator.ts
@@ -1,5 +1,3 @@
-import { BmiCategoryEnum } from "@entities/BmiCategoryEnum";
-
 export class BmiCalculator {
   public calculate(height: number, weight: number): number {
     return parseFloat((weight / Math.pow(height, 2)).toFixed(2));
