+++ tsconfig.json
@@ -35,7 +35,12 @@
       "@services/*": ["src/services/*"],
       "@test/*": ["src/test/*"],
       "@useCases/*": ["src/useCases/*"],
-      "@src/*": ["src/*"],
+      // "@src/*": ["src/*"],
+      "@enums/*": ["src/enums/*"],
+      "@interfaces/*": ["src/interfaces/*"],
+      "@utils/*": ["src/utils/*"],
+      "@dto/*": ["src/dto/*"],
+      "@controllers/*": ["src/controllers/*"],
     },                                      /* Specify a set of entries that re-map imports to additional lookup locations. */
     // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
     // "typeRoots": [],                                  /* Specify multiple folders that act like './node_modules/@types'. */
